refactor: extract overlays and configuration factory from ContentView
Moves document surface configuration builders, file drop handlers,
and scroll sync wiring into a ContentView extension in a separate
file. Extracts change navigation and watch pill overlays into named
computed properties. No inline overlay closures exceed 5 lines.

Part of #284
